Modernize cmake. Introduce project_options and project_warnings interface targets.

+++ src/basics/utilities/UbLogger.h
@@ -217,7 +217,7 @@ inline std::string UbLogger<OutputPolicy>::logTimeString()
     }
 
     char result[100]   = { 0 };
-    static DWORD first = GetTickCount64();
+    static DWORD first = (DWORD)GetTickCount64();
     std::sprintf(result, "%s.%03ld", buffer, (long)(GetTickCount64() - first) % 1000);
     return result;
 }
